requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytz>=2023.3
python-dotenv>=1.0.0
